랭크/일반 분기 처리

+++ Second_Project/src/main/java/org/joonzis/service/MemberService.java
@@ -12,7 +12,7 @@ public interface MemberService {
 	UserInfoDTO isValidUser(String user_id, String user_pw);
 	
 	// 유저 포인트 조회
-	int getUserPoint(String user_id);
+	long getUserPoint(String user_id);
 	
 	// 포인트 구매
 	void addPoint(String userId, int amount);
